Tracking: Prevent errors when double resuming sessions
When resuming sessions, an attempt to remove the file which holds the
sessions is made. If the sessions are resumed twice in succession then
two attempts are made but by the second attempt the file has already
been removed. Prevent errors by checking for the file's existence before
trying to remove it.

+++ kano_profile/tracker/tracking_sessions.py
@@ -267,7 +267,8 @@ def unpause_tracking_sessions():
     for session in get_paused_sessions():
         unpause_tracking_session(session)
 
-    os.remove(PAUSED_SESSIONS_FILE)
+    if os.path.exists(PAUSED_SESSIONS_FILE):
+        os.remove(PAUSED_SESSIONS_FILE)
 
 
 def get_session_event(session):
